requirements.txt
pygame-ce>=2.4.0
Pillow>=10.0.0
PyPDF2>=3.0.0
ebooklib>=0.18
lxml>=4.9.0
